# Tail-call interpreter dispatch