--- a/src/function.zig
+++ b/src/function.zig
@@ -6,6 +6,9 @@ const Opcode = @import("instruction.zig").Opcode;
 const Range = @import("common.zig").Range;
 
 pub const WasmError = error{
+    NotImplemented,
+    StackUnderflow,
+    StackOverflow,
     TrapUnreachable,
     LabelStackUnderflow,
     LabelStackOverflow,
@@ -16,6 +19,21 @@ pub const WasmError = error{
     BadFunctionIndex,
     ControlStackOverflow,
     BadInstanceIndex,
+    DivisionByZero,
+    Overflow,
+    InvalidConversion,
+    OutOfBoundsMemoryAccess,
+    IndirectCallTypeMismatch,
+    UndefinedElement,
+    //
+    BadMemoryIndex, // TODO: I think we won't see this with validation
+    MemoryIndexOutOfBounds, // TODO: I think we won't see this with validation?
+    BadTableIndex,
+    TableIndexOutOfBounds,
+    BadGlobalIndex,
+    GlobalIndexOutOfBounds,
+    NegativeDenominator,
+    Trap,
 };
 
 pub const Function = union(enum) {
@@ -48,26 +66,18 @@ pub const Instruction = union(Opcode) {
     block: struct {
         param_arity: usize,
         return_arity: usize,
-        // continuation: Range,
-        ip: usize,
-        ip_end: usize,
+        break_target: usize,
     },
     loop: struct {
         param_arity: usize,
         return_arity: usize,
-        // continuation: Range,
-        ip: usize,
-        ip_end: usize,
+        break_target: usize,
     },
     @"if": struct {
         param_arity: usize,
         return_arity: usize,
-        // continuation: Range,
-        // else_continuation: ?Range,
-        ip: usize,
-        ip_end: usize,
-        else_ip: usize,
-        else_ip_end: usize,
+        break_target: usize,
+        else_ip: ?usize,
     },
     @"else": void,
     end: void,
@@ -326,10 +336,7 @@ pub fn calculateContinuations(parsed_code_offset: usize, code: []Instruction) !v
             .block => |*block_instr| {
                 const end_offset = try findEnd(code[offset..]);
 
-                const continuation = code[offset + end_offset + 1 ..];
-                // block_instr.continuation = Range{ .offset = parsed_code_offset + offset + end_offset + 1, .count = continuation.len };
-                block_instr.ip = parsed_code_offset + offset + end_offset + 1;
-                block_instr.ip_end = block_instr.ip + continuation.len;
+                block_instr.break_target = parsed_code_offset + offset + end_offset + 1;
             },
             .@"if" => |*if_instr| {
                 const end_offset = try findEnd(code[offset..]);
@@ -337,24 +344,14 @@ pub fn calculateContinuations(parsed_code_offset: usize, code: []Instruction) !v
 
                 if (optional_else_offset == null and if_instr.param_arity -% if_instr.return_arity != 0) return error.ValidatorElseBranchExpected;
 
-                const continuation = code[offset + end_offset + 1 ..];
-                // if_instr.continuation = Range{ .offset = parsed_code_offset + offset + end_offset + 1, .count = continuation.len };
-                if_instr.ip = parsed_code_offset + offset + end_offset + 1;
-                // std.debug.warn("if_instr.ip = {}\n", .{if_instr.ip});
-                if_instr.ip_end = if_instr.ip + continuation.len;
+                if_instr.break_target = parsed_code_offset + offset + end_offset + 1;
+
                 if (optional_else_offset) |else_offset| {
-                    // const else_continuation = code[offset + else_offset + 1 ..];
-                    // if_instr.else_continuation = Range{ .offset = parsed_code_offset + offset + else_offset + 1, .count = else_continuation.len };
                     if_instr.else_ip = parsed_code_offset + offset + else_offset + 1;
-                    if_instr.else_ip_end = if_instr.else_ip + continuation.len;
                 }
             },
             .loop => |*loop_instr| {
-                // const end_offset = try findEnd(code[offset..]);
-                const continuation = code[offset..];
-                // loop_instr.continuation = Range{ .offset = parsed_code_offset + offset, .count = continuation.len };
-                loop_instr.ip = parsed_code_offset + offset;
-                loop_instr.ip_end = loop_instr.ip + continuation.len;
+                loop_instr.break_target = parsed_code_offset + offset;
             },
             else => {},
         }